% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCH_functions.R
\name{FastKerFdr}
\alias{FastKerFdr}
\title{FastKerFdr}
\usage{
FastKerFdr(Pval, p0 = NULL, plotting = FALSE, NbKnot = 1e+05, tol = 1e-05)
}
\arguments{
\item{Pval}{a vector of p-values (corresponding to a p-value serie)}

\item{p0}{a priori proportion of H0 hypotheses}

\item{plotting}{boolean, should some diagnostic graphs be plotted. Default is FALSE.}

\item{NbKnot}{The (maximum) number of knot for the kde procedure. Default is 1e5}

\item{tol}{a tolerance value for convergence. Default is 1e-5}
}
\value{
A list of 3 objects. Object p0 is an estimate of the proportion of H0 hypotheses.,
tau is the vector of H1 posteriors.
f1 is a numeric vector, each coordinate i corresponding to the evaluation of the H1 density at point pi, where pi is the ith p-value in Pval.
}
\description{
FastKerFdr
}
