% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fHconfig_sum_update_ptr_parallel}
\alias{fHconfig_sum_update_ptr_parallel}
\title{Computation of the sum sum_c(w_c*psi_c) parallelized version}
\usage{
fHconfig_sum_update_ptr_parallel(
  Hconfig,
  NewPrior,
  Logf0Mat,
  Logf1Mat,
  threads_nb = 0L
)
}
\arguments{
\item{Hconfig}{list of  vector of 0 and 1, corresponding to the configurations}

\item{NewPrior}{a double vector containing the prior w_c}

\item{Logf0Mat}{a double matrix containing the log(f0(xi_q))}

\item{Logf1Mat}{a double matrix containing the log(f1(xi_q))}

\item{threads_nb}{an int the number of threads}
}
\value{
a double vector containing sum_c(w_c*psi_c)
}
\description{
Computation of the sum sum_c(w_c*psi_c) parallelized version
}
