% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{gaussian_copula_density}
\alias{gaussian_copula_density}
\title{Gaussian copula density}
\usage{
gaussian_copula_density(zeta, R, Rinv)
}
\arguments{
\item{zeta}{the matrix of probit-transformed observations.}

\item{R}{the correlation matrix.}

\item{Rinv}{the inverse correlation matrix.}
}
\value{
A numeric vector, each coordinate \eqn{i} corresponding to the evaluation of the Gaussian copula density function at observation \eqn{\code{zeta}_i}.
}
\description{
Gaussian copula density
}
