% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHinfo.R
\name{GetH1AtLeast}
\alias{GetH1AtLeast}
\title{Specify the configurations corresponding to the composite H1.}
\usage{
GetH1AtLeast(Hconfig, AtLeast, Consecutive = FALSE, SameSign = FALSE)
}
\arguments{
\item{Hconfig}{A list of all possible combination of H0 and H1 hypotheses generated by the \code{\link[=GetHconfig]{GetHconfig()}} function.}

\item{AtLeast}{How many H1 hypotheses at least for the item to be of interest ? (an integer or a vector).}

\item{Consecutive}{Should the significant test series be consecutive ? (optional, default is FALSE).}

\item{SameSign}{Should the significant test series have the same sign ? (optional, default is FALSE).}
}
\value{
A vector 'Hconfig.H1' of components of Hconfig that correspond to the 'AtLeast' specification.
}
\description{
Specify the configurations corresponding to the composite H1.
}
\examples{
GetH1AtLeast(GetHconfig(4),2)

}
\seealso{
\code{\link[=GetH1Equal]{GetH1Equal()}}
}
