% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHinfo.R
\name{GetH1Equal}
\alias{GetH1Equal}
\title{Specify the configurations corresponding to the composite H1.}
\usage{
GetH1Equal(Hconfig, Equal, Consecutive = FALSE, SameSign = FALSE)
}
\arguments{
\item{Hconfig}{A list of all possible combination of H0 and H1 hypotheses generated by the \code{\link[=GetHconfig]{GetHconfig()}} function.}

\item{Equal}{What is the exact number of H1 hypotheses for the item to be of interest? (an integer or a vector).}

\item{Consecutive}{Should the significant test series be consecutive ? (optional, default is FALSE).}

\item{SameSign}{Should the significant test series have the same sign ? (optional, default is FALSE).}
}
\value{
A vector 'Hconfig.H1' of components of Hconfig that correspond to the 'Equal' specification.
}
\description{
Specify the configurations corresponding to the composite H1.
}
\examples{
GetH1Equal(GetHconfig(4),2)

}
\seealso{
\code{\link[=GetH1AtLeast]{GetH1AtLeast()}}
}
