% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHinfo.R
\name{GetHconfig}
\alias{GetHconfig}
\title{Generate the \eqn{H_0}/\eqn{H_1} configurations.}
\usage{
GetHconfig(Q, Signed = FALSE)
}
\arguments{
\item{Q}{The number of test series to be combined.}

\item{Signed}{Should the sign of the effect be taken into account? (optional, default is \code{FALSE}).}
}
\value{
A list '\code{Hconfig}' of all possible combination of \eqn{H_0} and \eqn{H_1} hypotheses among \eqn{Q} hypotheses tested.
}
\description{
Generate all possible combination of simple hypotheses \eqn{H_0}/\eqn{H_1}.
}
\examples{
GetHconfig(4)

}
