% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{last_incomplete_trapezoid_arma}
\alias{last_incomplete_trapezoid_arma}
\title{This function is a re-implementation of the initial R loop computing
last incomplete trapezoid. See R function integral.kde_adapted().}
\usage{
last_incomplete_trapezoid_arma(
  q_prob,
  q_ind,
  q,
  eval,
  est,
  simp_rule,
  density = TRUE
)
}
\arguments{
\item{q_prob}{reference of the vector q_prob containing the probability of each quantile q}

\item{q_ind}{reference of the vector q_ind}

\item{q}{reference of the vector q containing the quantiles}

\item{eval}{reference of the vector eval}

\item{est}{reference of the vector est}

\item{simp_rule}{reference of the vector simp_rule}

\item{density}{logical}
}
\value{
void. Its first argument q_prob is passed as a reference and modified in place.
}
\description{
This function is a re-implementation of the initial R loop computing
last incomplete trapezoid. See R function integral.kde_adapted().
}
