% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prior_update_arma_ptr_parallel}
\alias{prior_update_arma_ptr_parallel}
\title{Update of the prior estimate in EM algo parallelized version}
\usage{
prior_update_arma_ptr_parallel(
  Hconfig,
  fHconfig_sum,
  OldPrior,
  Logf0Mat,
  Logf1Mat,
  threads_nb = 0L
)
}
\arguments{
\item{Hconfig}{list of  vector of 0 and 1, corresponding to the configurations}

\item{fHconfig_sum}{a double vector containing sum_c(w_c*psi_c), obtained by fHconfig_sum_update_ptr_parallel()}

\item{OldPrior}{a double vector containing the prior w_c}

\item{Logf0Mat}{a double matrix containing the log(f0(xi_q))}

\item{Logf1Mat}{a double matrix containing the log(f1(xi_q))}

\item{threads_nb}{an int the number of threads}
}
\value{
a double vector containing the new estimate of prior w_c
}
\description{
Update of the prior estimate in EM algo parallelized version
}
