% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{PvalSets}
\alias{PvalSets}
\title{Synthetic example to illustrate the main qch functions}
\format{
A data.frame
}
\usage{
PvalSets
}
\description{
PvalSets is a data.frame with 10,000 rows and 3 columns. Each row corresponds to an item,
columns 'Pval1' and 'Pval2'  each correspond to a test serie over the items, and column 'Class'
provides the truth, i.e. if item \eqn{i} belongs to class 1 then the H0 hypothesis is true for the 2 tests,
if item \eqn{i} belongs to class 2 (resp. 3) then the H0 hypothesis is true for the first (resp. second)
test only, and if item \eqn{i} belongs to class 4 then both H0 hypotheses are false (for the first
and the second test).
}
\keyword{datasets}
