% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{PvalSets_cor}
\alias{PvalSets_cor}
\title{Synthetic example to illustrate the main qch functions using Gaussian copula}
\format{
A data.frame
}
\usage{
PvalSets_cor
}
\description{
PvalSets_cor is a data.frame with 10,000 rows and 3 columns. Each row corresponds to an item,
columns \code{Pval1} and \code{Pval2}  each correspond to a test serie over the items, and column 'Class'
provides the truth, i.e. if item \eqn{i} belongs to class 1 then the \eqn{H_0} hypothesis is true for the 2 tests,
if item \eqn{i} belongs to class 2 (resp. 3) then the \eqn{H_0} hypothesis is true for the first (resp. second)
test only, and if item \eqn{i} belongs to class 4 then both H0 hypotheses are false (for the first
and the second test). The correlation between the two pvalues series within each class is 0.3.
}
\keyword{datasets}
