% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{R.MLE.check}
\alias{R.MLE.check}
\title{Check the Gaussian copula correlation matrix Maximum Likelihood estimator}
\usage{
R.MLE.check(R)
}
\arguments{
\item{R}{Estimate of the correlation matrix.}
}
\value{
Estimate of the correlation matrix.
}
\description{
Check the Gaussian copula correlation matrix Maximum Likelihood estimator
}
