% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{R.MLE}
\alias{R.MLE}
\title{Gaussian copula correlation matrix Maximum Likelihood estimator.}
\usage{
R.MLE(Hconfig, zeta0, zeta1, Tau)
}
\arguments{
\item{Hconfig}{A list of all possible combination of H0 and H1 hypotheses generated by the \code{\link[=GetHconfig]{GetHconfig()}} function.}

\item{zeta0}{a matrix containing the Phi(F_0(Z_iq)), each column corresponding to a p-value serie.}

\item{zeta1}{a matrix containing the Phi(F_1(Z_iq)), each column corresponding to a p-value serie.}

\item{Tau}{a matrix providing for each item (in row) its posterior probability to belong to each of the H-configurations (in columns).}
}
\value{
Estimate of the correlation matrix.
}
\description{
Gaussian copula correlation matrix Maximum Likelihood estimator.
}
