% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{R_MLE_update_gaussian_copula_ptr_parallel}
\alias{R_MLE_update_gaussian_copula_ptr_parallel}
\title{Update the estimate of R correlation matrix of the gaussian copula, parallelized version}
\usage{
R_MLE_update_gaussian_copula_ptr_parallel(
  Hconfig,
  fHconfig_sum,
  OldPrior,
  Logf0Mat,
  Logf1Mat,
  zeta0,
  zeta1,
  OldR,
  OldRinv,
  RhoIndex,
  threads_nb = 0L
)
}
\arguments{
\item{Hconfig}{list of vector of 0 and 1, corresponding to the configurations}

\item{fHconfig_sum}{a double vector containing sum_c(w_c*psi_c), obtained by fHconfig_sum_update_ptr_parallel()}

\item{OldPrior}{a double vector containing the prior w_c}

\item{Logf0Mat}{a double matrix containing the log(f0(xi_q))}

\item{Logf1Mat}{a double matrix containing the log(f1(xi_q))}

\item{zeta0}{a double matrix containing the qnorm(F0(x_iq))}

\item{zeta1}{a double matrix containing the qnorm(F1(x_iq))}

\item{OldR}{a double matrix corresponding to the copula parameter}

\item{OldRinv}{a double matrix corresponding to the inverse copula parameter}

\item{RhoIndex}{a int matrix containing the index of lower triangular part of a matrix}

\item{threads_nb}{an int the number of threads}
}
\value{
a double vector containing the lower triangular part of the MLE of R
}
\description{
Update the estimate of R correlation matrix of the gaussian copula, parallelized version
}
