% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_decreasing_values_cpp}
\alias{remove_decreasing_values_cpp}
\title{Same as function above but does not handle the index ordering of the vector q.
Therefore, the 2nd argument order_q has to be an index ordered version of the vector q.
Indeed, the R base function: order() is twice as fast as the arma::sort_index(q)
This is therefore the recommended function to use.}
\usage{
remove_decreasing_values_cpp(q_prob, order_q, tol = 1e-10)
}
\arguments{
\item{q_prob}{reference of the vector q_prob}

\item{order_q}{reference of the vector q}

\item{tol}{By default 1e-10}
}
\value{
void. Its first argument q_prob is passed as a reference and modified in place.
}
\description{
Same as function above but does not handle the index ordering of the vector q.
Therefore, the 2nd argument order_q has to be an index ordered version of the vector q.
Indeed, the R base function: order() is twice as fast as the arma::sort_index(q)
This is therefore the recommended function to use.
}
