% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_decreasing_values_cpp_slow_ordering}
\alias{remove_decreasing_values_cpp_slow_ordering}
\title{This function is a re-implementation of the initial R side while loop.
See the end of R function integral.kde_adapted().
As shown in the commentary below, it is twice as slow to handle the index ordering
of the vector q (2nd argument) here with the function arma::sort_index().
Consequently, it is recommended to use the function remove_decreasing_values_cpp() instead.}
\usage{
remove_decreasing_values_cpp_slow_ordering(q_prob, q, tol = 1e-10)
}
\arguments{
\item{q_prob}{reference of the vector q_prob}

\item{q}{reference of the vector q}

\item{tol}{By default 1e-10}
}
\value{
void. Its first argument q_prob is passed as a reference and modified in place.
}
\description{
This function is a re-implementation of the initial R side while loop.
See the end of R function integral.kde_adapted().
As shown in the commentary below, it is twice as slow to handle the index ordering
of the vector q (2nd argument) here with the function arma::sort_index().
Consequently, it is recommended to use the function remove_decreasing_values_cpp() instead.
}
