% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{Copula.Hconfig_gaussian_density}
\alias{Copula.Hconfig_gaussian_density}
\title{Gaussian copula density for each H-configuration.}
\usage{
Copula.Hconfig_gaussian_density(Hconfig, F0Mat, F1Mat, R)
}
\arguments{
\item{Hconfig}{A list of all possible combination of \eqn{H_0} and \eqn{H_1} hypotheses generated by the \link[=GetHconfig]{GetHconfig()} function.}

\item{F0Mat}{a matrix containing the evaluation of the marginal cdf under \eqn{H_0} at each items, each column corresponding to a p-value serie.}

\item{F1Mat}{a matrix containing the evaluation of the marginal cdf under \eqn{H_1} at each items, each column corresponding to a p-value serie.}

\item{R}{the correlation matrix.}
}
\value{
A matrix containing the evaluation of the Gaussian density function for each H-configuration in columns.
}
\description{
Gaussian copula density for each H-configuration.
}
