% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{EM_calibration_indep_memory}
\alias{EM_calibration_indep_memory}
\title{EM calibration in the case of conditional independence with memory management (unsigned)}
\usage{
EM_calibration_indep_memory(
  Logf0Mat,
  Logf1Mat,
  Prior.init,
  Hconfig,
  Precision = 1e-06,
  threads_nb
)
}
\arguments{
\item{Logf0Mat}{a matrix containing the \eqn{\log(f_0(x^i_q))}}

\item{Logf1Mat}{a matrix containing the \eqn{\log(f_1(x^i_q))}}

\item{Prior.init}{the initialization of prior probabilities for each of the H-configurations.}

\item{Hconfig}{A list of all possible combination of \eqn{H_0} and \eqn{H_1} hypotheses generated by the \code{\link[=GetHconfig]{GetHconfig()}} function.}

\item{Precision}{Precision for the stop criterion. (Default is 1e-6)}

\item{threads_nb}{The number of threads to use.}
}
\value{
a vector of estimated prior probabilities for each of the H-configurations.
}
\description{
EM calibration in the case of conditional independence with memory management (unsigned)
}
