% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHinfo.R
\name{GetHconfig}
\alias{GetHconfig}
\title{Generate H0/H1 configurations.}
\usage{
GetHconfig(Q, Signed = FALSE)
}
\arguments{
\item{Q}{The number of test series to be combined.}

\item{Signed}{Should the sign of the effect be taken into account? (optional, default is FALSE).}
}
\value{
A list 'Hconfig' of all possible combination of H0 and H1 hypotheses among Q hypotheses tested.
}
\description{
Generate H0/H1 configurations.
}
\examples{
GetHconfig(4)

}
