% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qch-package.R
\docType{package}
\name{qch-package}
\alias{qch}
\alias{qch-package}
\title{qch: Query Composite Hypotheses}
\description{
Provides functions for the joint analysis of K sets of p-values obtained for a same list of items. This joint analysis is performed by querying a composite hypothesis, i.e. an arbitrary complex combination of simple hypotheses, as described in Mary-Huard et al. (2021) \href{https://arxiv.org/abs/2104.14601}{arXiv:2104.14601}. The null distribution corresponding to the composite hypothesis of interest is obtained by fitting non-parametric mixtures models (one for each of the simple hypothesis of the complex combination). Type I error rate control is achieved through Bayesian False Discovery Rate control. The 3 main functions of the package GetHinfo(), qch.fit() and qch.test() correspond to the 3 steps for querying a composite hypothesis (composed H0/H1 formulation, inferring the null distribution and testing the null hypothesis).
}
\author{
\strong{Maintainer}: Tristan Mary-Huard \email{tristan.mary-huard@agroparistech.fr} (\href{https://orcid.org/0000-0002-3839-9067}{ORCID})

Authors:
\itemize{
  \item Annaig De walsche \email{annaig.de-walsche@inrae.fr} (\href{https://orcid.org/0000-0003-0603-1716 }{ORCID})
}

}
\keyword{internal}
