% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qch_fit.R
\name{qch.fit}
\alias{qch.fit}
\title{Infer Hconfig posteriors}
\usage{
qch.fit(
  pValMat,
  EffectMat = NULL,
  Hconfig,
  copula = "indep",
  threads_nb = 0,
  plotting = FALSE,
  Precision = 1e-06
)
}
\arguments{
\item{pValMat}{A matrix of p-values, each column corresponding to a p-value serie.}

\item{EffectMat}{A matrix of estimated effects corresponding to the p-values contained in pValMat. If specified, the procedure will account for the direction of the effect. (optional, default is NULL)}

\item{Hconfig}{A list of all possible combination of H0 and H1 hypotheses generated by the \code{\link[=GetHconfig]{GetHconfig()}} function.}

\item{copula}{A string specifying the form of copula to use. Possible values are "indep"and "gaussian". Default is "indep" corresponding to the independent case.}

\item{threads_nb}{The number of threads to use. The number of thread will set to the number of core available by default.}

\item{plotting}{A boolean. Should some diagnostic graphs be plotted ? Default is FALSE.}

\item{Precision}{The precision for EM algorithm to infer the parameters. Default is 1e-6.}
}
\value{
If the storage permits, a list containing:
Object 'prior' is a vector of estimated prior probabilities for each of the H-configurations.
Object 'posterior' is a matrix providing for each item (in row) its posterior probability to belong to each of the H-configurations (in columns).
Object 'fHconfig' is a matrix containing config densities evaluated at each items, each column corresponding to a configurations.
Object 'Rcopula' is the estimated correlation matrix of the gaussian copula.(if applicable)
Object 'Hconfig' is the list of all configurations.
Else, a list containing:
Object 'prior' is a vector of estimated prior probabilities for each of the H-configurations.
Object 'f0Mat' is a matrix containing the evaluation of the marginal densities under H0 at each items, each column corresponding to a p-value serie.
Object 'f1Mat' is a matrix containing the evaluation of the marginal densities under H1 at each items, each column corresponding to a p-value serie.
Object 'F0Mat' is a matrix containing the evaluation of the marginal cdf under H0 at each items, each column corresponding to a p-value serie.
Object 'F1Mat' is a matrix containing the evaluation of the marginal cdf under H1 at each items, each column corresponding to a p-value serie.
Object 'fHconfig_sum' is a vector containing the sum over the configurations c of (w_c*psi_c) for each items.
Object 'Rcopula' is the estimated correlation matrix of the gaussian copula.(if applicable)
Object 'Hconfig' is the list of all configurations.
}
\description{
Infer Hconfig posteriors
}
\examples{
data(PvalSets)
PvalMat <- as.matrix(PvalSets[,-3])
## Build the Hconfig objects
Q <- 2
Hconfig <- GetHconfig(Q)

## Run the function
res.fit <- qch.fit(pValMat = PvalMat,Hconfig = Hconfig)

## Display the prior of each class of items
res.fit$prior

## Display the first posteriors
head(res.fit$posterior)
}
