% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qch_test.R
\name{qch.test}
\alias{qch.test}
\title{Perform composite hypothesis testing with FDR control.}
\usage{
qch.test(res.qch.fit, Hconfig, Hconfig.H1 = NULL, Alpha = 0.05)
}
\arguments{
\item{res.qch.fit}{The result provided by the \code{\link[=qch.fit]{qch.fit()}} function.}

\item{Hconfig}{A list of all possible combination of H0 and H1 hypotheses generated by the \code{\link[=GetHconfig]{GetHconfig()}} function.}

\item{Hconfig.H1}{An integer vector (or a list of such vector) of the Hconfig index corresponding to the H1 configuration(s). Can be generated by the \code{\link[=GetH1AtLeast]{GetH1AtLeast()}} or \code{\link[=GetH1Equal]{GetH1Equal()}} functions.
If NULL, the tests of being associated with "at least q" number of traits/studies, for q=1,..Q are performed.}

\item{Alpha}{the nominal Type I error rate for FDR control. Default is 0.05.}
}
\value{
A list of 3 objects 'Rejection', 'lFDR' and 'Pvalues'.
Object 'Rejection' is a matrix providing for each item the result of the composite hypothesis test, after multiple testing correction.
Object 'lFDR' is a matrix providing for each item its local FDR estimate.
Object 'Pvalues' a matrix providing for each item its pvalue of the composite hypothesis test.
}
\description{
Perform any composite hypothesis test by specifying the non-null H-configurations.
By default, perform the composite hypothesis testing of being associated with "at least q" number of traits/studies, for q=1,..Q.
}
\examples{
data(PvalSets)
PvalMat <- as.matrix(PvalSets[,-3])
Truth <- PvalSets[,3]

## Build the Hconfig objects
Q <- 2
Hconfig <- GetHconfig(Q)

## Infer the posteriors
res.fit <- qch.fit(pValMat = PvalMat, Hconfig = Hconfig)

## Run the test procedure with FDR control
H1config <- GetH1AtLeast(Hconfig,2)
res.test <- qch.test(res.qch.fit = res.fit,Hconfig = Hconfig, Hconfig.H1 = H1config)
table(res.test$Rejection$AtLeast_2,Truth==4)
}
\seealso{
\code{\link[=qch.fit]{qch.fit()}}, \code{\link[=GetH1AtLeast]{GetH1AtLeast()}},\code{\link[=GetH1Equal]{GetH1Equal()}}
}
